+++ src/main/idea/com/github/linwancen/plugin/graph/parser/groovy/GroovyParser.kt
@@ -26,8 +26,9 @@ class GroovyParser : ParserLang<GrMethod>() {
         return JavaParserUtils.skipFun(state, func)
     }
 
-    override fun toSign(func: GrMethod): String {
-        return "${func.containingClass?.qualifiedName ?: ""}#${func.name}"
+    override fun toSign(func: GrMethod): String? {
+        val clazz = func.containingClass ?: return "#${func.name}"
+        return "${clazz.qualifiedName ?: return null}#${func.name}"
     }
 
     override fun funMap(funMap: MutableMap<String, String>, func: GrMethod) {